+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.3 (SNAPSHOT)
-*
+* Modify cli readConfigFileSync, config file in current directory should take precedence over the one in home directory
 
 ### 0.0.2
 * Add mock http request and response